feat(http_api): got an empty http seriver running
A few things I learnt along the way:

1. sbt expects the project to be in a very specific directory strucutre.
Namely, it needs to be like:

```
src/
  main/
    scala/
      top-domain/
        second-level-domain/
          sub-domain/
            ...
    resources/
      ...
```

To me, this implicitness feels kinda dirty. Especially for the case with
pureconfig automatically looking in the resources/ directory for a
specific application.conf file. It would have been a bit clearer if this
was more explicit. Perhaps the more I use Scala, the more this will
become normal.

2. I've also added a basic pureconfig Config class to pull environment
variables into the app

3. Added a basic Dockerfile to build and run the app

4. Updated build.sbt to include http4s