refactor(http_api): make code more scala-idiomatic
1. Try to stay within ConnectionIO land as long as possible (so we're
still within a single transaction). So I've updated
CreateMessageRepository.create to not execute operation but to just
return a ConnectionIO[Message].
2. Repository should accept a transactor of a higher class
(HikariTransactor extends Transctor) - more generic
3. Clean up the MessageService.createMessage function. A few things
here:

- Try to treat everything as values (everything becomes a lot easier)
- Continue to flatMap on create(stagedMessage) but store that as a
"value" I can flatmap/map later.
- ^ works because we're still in ConnectionIO land.

When we flatMap(sendMessage) our ConnectionIO, we need to return a
ConnectionIO but we want to get the message we created. Issue is we also
need to send to a queue and ensure we're still in the same transaction.

So to solve this, there's a function `.delay` we can use which I believe is
similar to Task.xxx exception instead of returning a Task, it's a
ConnectionIO. By chaining these ConnectionIO "values? functions?" together,
we can stay within ConnectionIO land and can avoid executing until we
need to (I think).

Previously if our queue.send fails then we won't get a database rollback
because we left ConnectionIO land and hence left the transaction.
However, we only exit ConnectionIO land after we perform a
queue.send (repo.executeOp is called after we flatMap createdMessageOp)